requirements.txt
moondream>=0.0.5
opencv-python>=4.9.0
Pillow>=10.0.0
pyyaml>=6.0
numpy>=1.24.0
reachy-mini-sdk
